Remove server and shared sprite component
