+++ Robust.Shared/Serialization/TypeSerializers/Implementations/ResPathSerializer.cs
@@ -28,7 +28,7 @@ public sealed class ResPathSerializer : ITypeSerializer<ResPath, ValueDataNode>,
 
         if (!path.CanonPath.Split('/').First().Equals("Textures", StringComparison.InvariantCultureIgnoreCase))
         {
-            path = SharedSpriteComponent.TextureRoot / path;
+            path = SpriteSpecifierSerializer.TextureRoot / path;
         }
 
         path = path.ToRootedPath();
